include canonicalization in the cache key
Prior to this patch, since the cache manager didn't include the
canonicalization flag when computing the cache key, Wave occasionally
retrieved an incorrect kernel from the cache.  Specifically, if the
cached kernel was not run through canonicalization (since the flag is
off by default) and if the programmer requested canonicalization, the
programmer would still get the non-canonicalized (but cached) IR.

This patch fixes the problem by including canonicalization among the
details that are included when computing the cache key.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>