compute cache key using all options that may affect compilation
Several other options besides canonicalization may impact the generated
IR and these were not included in the computation of the cache key,
which would have the effect of retrieving an incorrect IR from the cache
when these options changed from one compilation to the next.  This patch
adds these new options to the function that computes the cache key.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>