[Wave] Teach compiler to do masking properly on scaled dims.
Currently scaled dimensions such as K/32 or K/2 only works on aligned
shape(wrt block_k). In order to add masking/unaligned support for this
we add feature to check that we have indeed scaled bound and try to
scale the bound. This is most useful for handling K dims that are not
power of 2 for Llama3

Signed-off-by: Stanley Winata <stanley.winata@amd.com>