[Wave] fix tutorial ref to tkl
The tkl import is actually unused in this example, so an alternative is to just remove it.  However, since going from a tutorial kernel to a custom kernel is going to be a common use-case, and you are going to reach for tkl, it may be helpful to have it (especially since other tutorials that reference tkl don't specify the actual import).  But if we have it it should be correct.

Signed-off-by: William G Hatch <william@hatch.uno>