add constraints to the cache key
The `anonymize_constraints()` function has a bug, in that it updates the
constraints as desired, but it doesn't return the updated constraints.
At the same time, in the caller, we use the return value of the function
to create the cache key.  Consequently, prior to this patch, we always
used `None` as the constraint portion of the cache key, so even after
updating the constraints, Wave still retrieved a stale compiled kernel
from the cache.

Additionally, prior to this change, the uncached compilation path
initialized/updated the constraints, thus causing a cache miss even
though the kernel and the programmer-specified constraints didn't
change.  Luckily, the updates to the constraints seem independent of the
trace, so this patch moves the constraint updates _before_ the cache
lookup.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>