NFC move some reordering logic for debug logs
This will be helpful when there is more processing.  I originally did this in https://github.com/iree-org/wave/pull/95 but I have since decided to approach the main aspect of it in a different way, I'm pulling out this small NFC that will still be helpful for other ways to write printing / visualizingg features.

Signed-off-by: William G Hatch <william@hatch.uno>