rename debug_log_write to debug_log
and its argument from log_name to label.

I originally wrote debug_log_write thinking it would be a primitive operation that I would write other operations on top of, but now I think we will just use this one op instead of making variants of it.  The old names are maybe more descriptive, but for its use as a debugging op, I think it will be worthwhile to make the names shorter and easier to write.

Since we just added this feature and haven't had a major release yet, and this is a debugging tool that is explicitly marked as not being stable and that should be removed from production kernels, I think it is fair to change the API like this.

Depends on https://github.com/iree-org/wave/pull/118

Signed-off-by: William G Hatch <william@hatch.uno>