[debugging] implement an html generation debug_log viewer
Use in a wave kernel like:

```
a_reg = tkw.read(...)
from wave_lang.debugging.html_viewer import html_viewer
tkw.debug_log(a_reg, handler=html_viewer)
```

This creates a `debug-view.html` file in the current directory.
Open the HTML file in a browser to view log outputs.

Scalar, 1d, and 2d data is viewed the way you expect.
3d or higher dimensional data is shown with certain dimensions fixed to a value (starting at 0), which you can change.
For all views, you can change which dimensions are fixed, or transpose data, by adjusting a view string, then hitting the `Update` button (or pressing enter in the text box).
An example view string for an tensor of symbolic shape `(A,B,C)` looks like `A=0, B, C`, but an alternate and transposed view could be `B=5, C, A`.

The tensors  by default are shown in a “Compressed Global” view, meaning that the full global tensor is shown, but in compressed view that shows at most 6 rows and 6 columns, like the Torch tensor print format.
There is a dropdown to change to other views, which for now only has “Full Global”.
In the future I plan to add a block-level view, a wave-level view, and a thread-level view.
Though I haven't yet decided what all of the UI around selecting which block/wave/thread to view will be like.

Signed-off-by: William G Hatch <william@hatch.uno>