implement TopkOp
This implements a TopkOp with a similar shape to the ReduceOp.

The algorithm it uses is similar to
https://github.com/sgl-project/sglang/blob/6337d9057c78fc29b31d15bec892bb9013201f6d/sgl-kernel/csrc/moe/moe_topk_softmax_kernels.cu#L185 -- it does a reduction to find the maximum value and index, then masks the value at that index and repeats.

There are some things cargo-culted from the implementation of ReduceOp.  I don't yet understand the placeholder stuff that is done when creating a conditional region during the rewriting.

This PR depends on https://github.com/iree-org/wave/pull/251

Signed-off-by: William G Hatch <william@hatch.uno>