Use string instead of index for updating the write dependency
Prior to this patch, adding new fields at the end of the ReadOp type
introduced problems because the code subtly changed the newly introduced
field instead of the write dependency even though it was asked to update
just the write dependency.  This patch fixes the problem by using the
name of the field instead of the index.