add location_check_pass
This adds a pass that makes no changes to the trace, but that logs information about debug locations.

Note that this PR doesn't actually add the pass to the pipeline, it just adds the pass so that it can be used.  In working on locations I'm using this pass, and I figure I should just check it in.

Signed-off-by: William G Hatch <william@hatch.uno>