Fix argument type for `amdgpu.fat_raw_buffer_cast` op
The commit hash of LLVM used by IREE does not yet include the upstream
change (2195fe7e) that updated the integer type for the `validBytes`
from i32 to i64.  So when Wave emits `amdgpu.fat_raw_buffer_cast` ops,
we get a compilation failure.

This patch fixes the type to match the LLVM commit used by IREE/Wave,
although in a few weeks when the upstream change is pulled into Wave,
we'll need to revert this patch.