Use str instead of sympy.Symbol to convert to
affine map

Original review: https://github.com/ftynse/water/pull/52

When converting sympy expression to affine map, we only care about the
symbol names and not the exact `sympy.Symbol` objects. This allows the
caller of `convert_sympy_to_affine_map` to freely add assumptions to the
symbols (hence creating a new symbol with the same name) and substitute
the original symbols in the expression.

Signed-off-by: tyb0807 <sontuan.vu@amd.com>