Add asm backend for compiler
This PR adds a new asm backend for the Wave compiler.
The asm backend lowers directly from MLIR to AMD GCN ISA.
This is a an experimental backend and currently we only
lower a copy example and add tests to validate correctness.
Since we do not generate VMFBs, this only works with the
wave runtime.

A new asm directory is added with support for instructions,
expression simplification. Documentation that describes
the current backend is also added.

Signed-off-by: Harsh Menon <harsh@nod-labs.com>