Add wave.schedule for more fine grained control (#333)
This PR adds a new language construct, the wave.schedule which gives
users more control over the generated code. Specifically, this PR allows
authors to control the pipelining and specify which operators or groups
of operators belong to which stage of the pipeline. A lit test and e2e
test are added. The new schedule is traced to a ScheduleRegionGraph
(distinct from RegionGraph which is only for the kernel) and a
corresponding ScheduleContext for tracing the schedule graph.

---------

Signed-off-by: Harsh Menon <harsh@nod-labs.com>
Co-authored-by: Harsh Menon <harsh.menon@amd.com>
Signed-off-by: sa-faizal <sahil.faizal@amd.com>