Rewrite more reads and writes with gather-to-lds operations
This patch makes three main changes.  First, it moves the
gather-to-shared passes to run after the minimize-global-loads pass.
Doing so uncovers new opportunities for rewriting read/write operations.
Second, this patch fixes a bug in the handling of dynamic values map
where we were incorrectly using the same map for both source and
destination values.  Third, this patch fixes a bug in the rewriting of
read and write index values, where instead of the `symbolic_shape`, we
should have used `indexing_dims`.

The net effect of these changes is that Wave discovers more
opportunities for optimizations in several attention kernels.