Add optional return values support for Conditional nodes
This commit the `conditional` operation to support optional return values, enabling conditionals to propagate data through if-else branches similar to how `iterate` operations handle loop-carried variables.

To return values, you use a `return` statement in the decorated function that forms the body of the conditional.
The conditional form doesn't have an `else` body, but if there is a return value for the true branch, there must also be a return for the false branch, so this commit adds the `init_args` optional argument to conditionals that is used for the default return.
The name `init_args` is used to maintain some API compatibility with the `iterate` operation, since many pieces of the compiler operate on both of them generically.

Signed-off-by: William G Hatch <william@hatch.uno>