Switch emitter to use only upstream dialects (#359)
Switch emitter to generate kernel body using only upstream dialects and
then generate IREE wrapper on top of it.

This is a preparation step for completely decoupling wave_runtime from
IREE.

The only significant change (and source of lit tests churn) is that
buffer arguments are now passed as 0D memrefs and then
`reinterpret_cast`d to their real shape. This is done to simplify future
kernel ABI and avoid dealing with memref descriptors on host<->GPU
boundaries.

---------

Signed-off-by: Ivan Butygin <ivan.butygin@gmail.com>
Signed-off-by: sa-faizal <sahil.faizal@amd.com>