[water] add iter symbol attribute
This mirrors the index symbol attribute and lets us differentiate induction
variables used in indexing expressions from regular symbols with the same name
without resorting to magic naming schemes in the C++ code. Since we are reusing
affine expression parser that expects a bare keyword, we still have to resort
to the magic naming scheme prefixing with `_Iter` in the textual format of the
IR. Harden the verifier to complain about regular symbol names starting with an
underscore and make them reserved.

Signed-off-by: Alex Zinenko <git@ozinenko.com>