[wave2water] drop irrelevant induction variable symbols
These should not appear in the expression to start with. I couldn't quickly
identify the source of this bug, it's deep inside index sequence propagation
logic that is not localized and is, e.g., mutating the constraints in the
process. This hasn't been found before because the index is simply ignored for
allocate operations and there don't seem to be other tests with a similar
structure.

Signed-off-by: Alex Zinenko <git@ozinenko.com>