Force garbage collection in e2e tests
This should free the GPU memory and avoid the situation where tests fail
because they cannot allocate anymore. There is some extra runtime penalty, but
we only pay it per module (file) and not per test invocation.

The approach uses the pytest 'plugin' approach based on having a file called
`conftest.py` in the directory.

Signed-off-by: Alex Zinenko <git@ozinenko.com>