Enable Water-based index inference as optional secondary
This starts the transition to use Water-based index expression inference by
enabling it to run in parallel with the original pyWave version of the same,
optionally under a flag. Only the pyWave analyses or both may run. When both
analyses run, the results of the two are compared with exceptions raised on
mismatches. In any case, the results of the pyWave analyses are taken.

After stabilization, appropraite extension and fixes, the following steps will
be taken:

1. Turn the flag running the Water-based analysis to be on by default.
2. Make the Water-based analysis primary and keep the pyWave analysis as
  optional backup for comparison.
3. Deprecate and remove the pyWave analysis.

Each of these steps will come in a separate commit after a certain period of
time.

The plumbing requires conversion from MLIR attributes to Wave constructs,
mostly sympy expressions that is added to the extent required here. The logic
in `emit_wave_dialect` is extended to collect attribute values for all
operations in the IR based on a unique ID derived from Python hash of the
object attached as attribute.

Some portion of Wave code interacting with sympy is shifted to
`wave_lang/support` to avoid the unconditional loading of IREE libraries due to
`wave_lang.kernel` including `wave_lang.kernel.wave` which, in one or multiple
places, transitively imports IREE, which in turn clashes with Water or any
other MLIR-based project.

One of the tests is added in a separate subdirectory and excluded from pytest
global collection as the collection process triggers transitive imports and
therefore imports IREE, which we don't want.

Signed-off-by: Alex Zinenko <git@ozinenko.com>